// src/types/app-page.ts
export type Page =
  | "home"
  | "dashboard"
  | "inventory"
  | "mediamap"
  | "clients"
  | "products"
  | "proposals"
  | "campaigns"
  | "reservations"
  | "financial"
  | "messages"
  | "marketplace-inquiries"
  | "marketplace-management"
  | "mediakit"
  | "promotions"
  | "activities"
  | "settings"
  | "superadmin";